allow larger diff in tests; exit os tests if error

+++ tests/tests_os.py
@@ -10,19 +10,21 @@ import nibabel as nib
 from totalsegmentator.python_api import totalsegmentator
 
 
-if __name__ == "__main__":
+def run_tests_and_exit_on_failure():
 
     # Test python api
     # Test organ predictions - fast
     totalsegmentator("tests/reference_files/example_ct_sm.nii.gz", "tests/unittest_prediction_fast", fast=True, device="cpu")
-    pytest.main(["-v", "tests/test_end_to_end.py::test_end_to_end::test_prediction_fast"])
+    r = pytest.main(["-v", "tests/test_end_to_end.py::test_end_to_end::test_prediction_fast"])
     shutil.rmtree("tests/unittest_prediction_fast")
+    if r != 0: sys.exit("Test failed: test_prediction_fast")
 
     # Test python api - nifti input
     input_img = nib.load("tests/reference_files/example_ct_sm.nii.gz")
     totalsegmentator(input_img, "tests/unittest_prediction_fast", fast=True, device="cpu")
-    pytest.main(["-v", "tests/test_end_to_end.py::test_end_to_end::test_prediction_fast"])
+    r = pytest.main(["-v", "tests/test_end_to_end.py::test_end_to_end::test_prediction_fast"])
     shutil.rmtree("tests/unittest_prediction_fast")
+    if r != 0: sys.exit("Test failed: test_prediction_fast with Nifti input")
 
     # Test terminal
     # Test organ predictions - fast - multilabel
@@ -31,10 +33,16 @@ if __name__ == "__main__":
     file_in = os.path.join("tests", "reference_files", "example_ct_sm.nii.gz")
     file_out = os.path.join("tests", "unittest_prediction_fast.nii.gz")
     subprocess.call(f"TotalSegmentator -i {file_in} -o {file_out} --fast --ml -d cpu", shell=True)
-    pytest.main(["-v", "tests/test_end_to_end.py::test_end_to_end::test_prediction_multilabel_fast"])
+    r = pytest.main(["-v", "tests/test_end_to_end.py::test_end_to_end::test_prediction_multilabel_fast"])
     os.remove(file_out)
+    if r != 0: sys.exit("Test failed: test_prediction_multilabel_fast")
 
     # Test Dicom input
     totalsegmentator("tests/reference_files/example_ct_dicom", "tests/unittest_prediction_dicom.nii.gz", fast=True, ml=True, device="cpu")
-    pytest.main(["-v", "tests/test_end_to_end.py::test_end_to_end::test_prediction_dicom"])
+    r = pytest.main(["-v", "tests/test_end_to_end.py::test_end_to_end::test_prediction_dicom"])
     os.remove("tests/unittest_prediction_dicom.nii.gz")
+    if r != 0: sys.exit("Test failed: test_prediction_dicom")
+
+
+if __name__ == "__main__":
+    run_tests_and_exit_on_failure()
